requirements.txt
requests==2.31.0
pathvalidate==3.1.0
python-dotenv==1.0.0
python-telegram-bot==13.13
urllib3==2.0.3